working on integrating chunk data between java and rust
