+++ fabric/src/main/java/dev/birb/wgpu/WgpuMcMod.java
@@ -1,14 +1,27 @@
 package dev.birb.wgpu;
 
+import dev.birb.wgpu.render.electrum.ElectrumRenderer;
+import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.renderer.v1.Renderer;
+import net.fabricmc.fabric.api.renderer.v1.RendererAccess;
+import net.fabricmc.fabric.impl.client.indigo.Indigo;
+import net.fabricmc.fabric.impl.client.indigo.IndigoMixinConfigPlugin;
+import net.fabricmc.fabric.impl.renderer.RendererAccessImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class WgpuMcMod implements ModInitializer {
+public class WgpuMcMod implements ClientModInitializer {
 
 	public static Logger LOGGER = LoggerFactory.getLogger("wgpu_mc");
 
+	public static ElectrumRenderer ELECTRUM;
+
 	@Override
-	public void onInitialize() {
+	public void onInitializeClient() {
+		LOGGER.info("[Electrum] Registering Electrum renderer!");
+
+		ElectrumRenderer electrumRenderer = new ElectrumRenderer();
+		RendererAccess.INSTANCE.registerRenderer(electrumRenderer);
 	}
 }
