+++ fabric/src/main/java/dev/birb/wgpu/render/electrum/ElectrumMeshBuilderImpl.java
@@ -1,0 +1,19 @@
+package dev.birb.wgpu.render.electrum;
+
+import net.fabricmc.fabric.api.renderer.v1.mesh.Mesh;
+import net.fabricmc.fabric.api.renderer.v1.mesh.MeshBuilder;
+import net.fabricmc.fabric.api.renderer.v1.mesh.QuadEmitter;
+
+public class ElectrumMeshBuilderImpl implements MeshBuilder {
+
+    @Override
+    public QuadEmitter getEmitter() {
+        return null;
+    }
+
+    @Override
+    public Mesh build() {
+        return null;
+    }
+
+}
