+++ fabric/src/main/resources/wgpu.mixins.json
@@ -4,32 +4,6 @@
   "package": "dev.birb.wgpu.mixin",
   "compatibilityLevel": "JAVA_8",
   "client": [
-    "TitleScreenMixin",
-    "accessors.DefaultClientResourcePackAccessor",
-    "accessors.NamespaceResourceManagerAccessor",
-    "accessors.ReloadableResourceManagerImplAccessor",
-    "accessors.ThreadExecutorAccessor",
-    "accessors.VertexFormatElementAccessor",
-    "accessors.WindowAccessor",
-    "core.ClientPlayNetworkHandlerMixin",
-    "core.GlStateManagerMixin",
-    "core.KeyboardMixin",
-    "core.MainMixin",
-    "core.MinecraftClientCoreMixin",
-    "core.MouseMixin",
-    "core.RegistryMixin",
-    "render.BufferRendererMixin",
-    "render.FramebufferMixin",
-    "render.GameRendererMixin",
-    "render.GLXMixin",
-    "render.MinecraftClientRenderMixin",
-    "render.RenderSystemMixin",
-    "render.WindowFramebufferMixin",
-    "render.WindowMixin",
-    "render.WorldRendererMixin",
-    "world.PaletteProviderMixin",
-    "world.WorldChunkMixin",
-    "world.BlockStateMixin"
   ],
   "injectors": {
     "defaultRequire": 1
